1.8b update
Note: changes of the update would be noticed in the app version of the browser.

this update is to get rid of the temporary way for the AAM to work using a weird and modified version of the normal W1.1 and instead now it depends on the actual  W1.1b update which would be the base method and not temporary.