THE 1.8GF\1.8G2 UPDATE (AKA 'Kakudai flowers full\Kakudai flowers2')
NOTE:
THIS UPDATE FOR THE APP VERSION IS CONSIDERED THE SECOND PART OF THE "THE 1.8G UPDATE (AKA 'Kakudai flowers')"

note: the first part release is exclusive to app version

BUT FOR THE CODE VERSION(NOT APP VERSION) THIS UPDATE IS THE FULL RELEASE OF BOTH UPDATES(1.8G and 1.8G2) IN ONE UPDATE(the two updates which got released separately for the app version are released here in one big update) and this big update which got both versions release at once and it's exclusive for the code version only named
(1.8GF,Kakudai flowers full)
 which got 2 main parts:

part1:

note: from this update(1.8G), only the updates that need changes of WBSE compared
to the version before it  would be adde5d as a new version in the W series, and only the update that need changes for their AMM compared to the version before it would be added as a new version in the A series.

this update is an improvement to the interface , firstly we minimized the required width and height that it become even less than Microsoft edge's minimum width and height, secondly we improved the homepage and incognito a lot like a lot we made them have different cases and made them working much better with smaller dimensions esp. width , finally we done some other GUI changes like removing the spaces between the 'new tab' button and 'extensions' button and even shortened the space between the 'refresh' button and URL and name bars plus we minimized the minimum bars width from 100 pixels to 20 pixels.

part2:

 this update(1.8G2) is for updating hard coded homepages in the system with the new changes
and fixing the 'Forgotten <' problem in normal homepage